+++ Project-Perfect-Citizen/Code/Game/bootLoadingAnimationRender.cpp
@@ -185,7 +185,7 @@ string bootLoadingAnimationRender::getRandomString(int stringLength) {
     std::string text = "";
     std::string possible = "ABCDEFGHIJKLMNOPQRST!@#$%^&*/_+=UVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
     for (int i = 0; i < stringLength; ++i) {
-        text += possible[floor(rand() % possible.length())];
+        text += possible[int(floor(rand() % possible.length()))];
     }
     return text;
 }
